wget-parts: remove SIG handlers
It seems that the commands which run with an ampersand '&' had better
be easily killable.  Otherwise, it is harder to kill the whole thing.
Unless the process is killed immediately, bash tries something like
waiting for the process and other funny things; parts of the code
run on after SIGINT.