wget-parts: do not permit empty *.ts files
I'm getting zero-sized *.ts parts, most probably due to a faulty proxy.
In fact, a *.ts file cannot be zero-size legitimately; also, since
is its content is static, the size should be readily available, unless
a proxy is just too rotten.  Thus it makes sense to check check for a
non-zero $right_size.

Update: unfortunately, some proxies out there are "rotten" in that they
don't set the static content's size; but some of them are pretty fast,
and sometimes they're the best we can hope for.  And so the check for
the non-zero $right_size has been relaxed again.